move check command to archiver.check
